core: Move the `FilesetParser` to it
So `jj-core` users also have access to it, since we already did the same thing for the `RevsetParser`.
This is part of building a new `jj-core` crate.

Part of #6284
